Create and refactor into separate test projects
