+++ WindowsFirewallNotifier.sln
@@ -16,7 +16,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestProject", "TestProject\TestProject.csproj", "{D6117BE3-3E69-48F9-90DC-B53AA29613A6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Console.Tests", "Console.Tests\Console.Tests.csproj", "{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common.Tests", "Common.Tests\Common.Tests.csproj", "{E6C00065-F1D9-407C-99D9-14DD3EECFF45}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -76,17 +78,30 @@ Global
 		{94028C5E-9B0E-4D15-966A-9FA766A20821}.Release|x64.Build.0 = Release|x64
 		{94028C5E-9B0E-4D15-966A-9FA766A20821}.Release|x86.ActiveCfg = Release|x86
 		{94028C5E-9B0E-4D15-966A-9FA766A20821}.Release|x86.Build.0 = Release|x86
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Debug|x64.Build.0 = Debug|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Debug|x86.Build.0 = Debug|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Release|x64.ActiveCfg = Release|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Release|x64.Build.0 = Release|Any CPU
-		{D6117BE3-3E69-48F9-90DC-B53AA29613A6}.Release|x86.ActiveCfg = Release|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Debug|x64.Build.0 = Debug|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Debug|x86.ActiveCfg = Debug|x86
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Debug|x86.Build.0 = Debug|x86
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Release|Any CPU.Build.0 = Release|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Release|x64.ActiveCfg = Release|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Release|x64.Build.0 = Release|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Release|x86.ActiveCfg = Release|Any CPU
+		{83A7C9E9-4035-4CEF-BCAB-FD19999C4619}.Release|x86.Build.0 = Release|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Debug|x64.Build.0 = Debug|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Debug|x86.ActiveCfg = Debug|x86
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Debug|x86.Build.0 = Debug|x86
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Release|x64.ActiveCfg = Release|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Release|x64.Build.0 = Release|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Release|x86.ActiveCfg = Release|Any CPU
+		{E6C00065-F1D9-407C-99D9-14DD3EECFF45}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
